# Optimize small-limb multiplication path