## Summary
- factor out 128-bit and 256-bit multiplication helpers to streamline operator* and retain fast path for small operands
- use shared helpers for limb-by-limb multiplication

## Testing
- `cmake --build build`
- `cd build && ctest --output-on-failure`
- `./build/perf_compare_int256_cxx11 --benchmark_filter="Mul/(Small|Large|Mixed)/Wide" --benchmark_format=json`
  - Before: Small 3.06 ns; Large 5.42 ns; Mixed 5.50 ns
  - After: Small 2.90 ns; Large 5.34 ns; Mixed 5.46 ns

------
https://chatgpt.com/codex/tasks/task_e_68a8269adbac832997d27e64a10cb480